chatkit-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=2.8.1
pydantic>=2.9.0
openai-chatkit>=1.4.0
openai-agents>=0.6.1
